+++ packages/lodash-pro/scripts/build.ts
@@ -1,6 +1,6 @@
 import { build } from '~scripts/build'
-import tsupConfig from './tsup.config'
+import { cjs, esm } from './tsup.config'
 
-build(tsupConfig('cjs'))
+build(esm)
 
-build(tsupConfig('esm'))
+build(cjs)