+++ packages/postcss-config/scripts/tsup.config.ts
@@ -2,6 +2,6 @@ import { getEntry } from '~scripts/utils'
 import { type Options } from 'tsup'
 
 export default {
-  entry: await getEntry('src/**/index.ts'),
+  entry: getEntry('src/**/index.ts'),
   target: 'esnext',
 } as Options