+++ packages/react-component/scripts/build.ts
@@ -1,11 +1,5 @@
 import { build } from '~scripts/build'
-import { getEntry } from '~scripts/utils'
-import tsupConfig from './tsup.config'
+import { cjs, esm } from './tsup.config'
 
-build({
-  ...tsupConfig,
-  format: ['esm', 'cjs'],
-  dts: {
-    entry: await getEntry('src/**/*.{ts,tsx}'),
-  },
-})
+build(cjs)
+build(esm)