+++ packages/react-component/scripts/dev.ts
@@ -1,8 +1,4 @@
 import { dev } from '~scripts/dev'
-import tsupConfig from './tsup.config'
+import { esm } from './tsup.config'
 
-dev({
-  ...tsupConfig,
-  format: ['esm'],
-  dts: false,
-})
+dev(esm)