+++ packages/react-hook/scripts/dev.ts
@@ -1,4 +1,6 @@
 import { dev } from '~scripts/dev'
-import tsupConfig from './tsup.config'
+import { cjs, esm } from './tsup.config'
 
-dev(tsupConfig)
+dev(cjs)
+
+dev(esm)