+++ packages/vite-config/src/node/plugins/svgr.ts
@@ -1,5 +1,5 @@
-import svgrPlugin, { type ViteSvgrOptions } from 'vite-plugin-svgr'
+import svgrPlugin, { type VitePluginSvgrOptions } from 'vite-plugin-svgr'
 
-export function svgr(options: ViteSvgrOptions) {
+export function svgr(options: VitePluginSvgrOptions) {
   return svgrPlugin(options)
 }