+++ packages/react-component/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @minko-fe/react-component
 
+## 0.1.20
+
+### Patch Changes
+
+- better tree-shaking
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.2
+  - @minko-fe/react-hook@0.2.2
+
 ## 0.1.19
 
 ### Patch Changes