+++ packages/react-hook/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @minko-fe/react-hook
 
+## 0.2.2
+
+### Patch Changes
+
+- better tree-shaking
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes